backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
web3==6.11.3
eth-account==0.9.0
python-dotenv==1.0.0
redis==5.0.1
openai==1.3.7
numpy==1.24.3
stripe==7.8.0 